scheduler_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.27.0
google-api-python-client==2.118.0
redis==5.0.1
sendgrid==6.12.0
pytz==2024.1
alembic==1.13.1
pydantic==2.7.1  # Critical update